requirements.txt
asgiref==3.2.10
Django==3.0.8
django-ckeditor==5.9.0
django-debug-toolbar==2.2
django-js-asset==1.2.2
Pillow==7.2.0
pkg-resources==0.0.0
psycopg2==2.8.5
python-decouple==3.3
pytz==2020.1
sqlparse==0.3.1
